Fix: forgot to include limit & moved includes.hpp
