+++ Src/Client/Input/userInput.hpp
@@ -1,7 +1,6 @@
 #ifndef USER_INPUT_HPP
 #define USER_INPUT_HPP
-#include "../includes.hpp"
-using namespace std;
+#include "../includes/includes.hpp"
 
 int input = 0;
 bool input_exception = false;
