+++ Src/Client/main.cpp
@@ -1,4 +1,4 @@
-#include "../includes.hpp"
+#include "./includes/includes.hpp"
 #include "Display/display.hpp"
 
 int main()
